eliminate shared etcd for integration tests
